// generated/src/app/api-models/cap-stoploss.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapStoploss {
  capModelId: string;
  capStoplossId: string;
  description: string;
  amountTypeSl: string;
  stoplossThreshold: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}